+++ index.php
@@ -1,4 +1,7 @@
 <?php
+
+@ini_set("memory_limit",'128M');
+
 /**
  * Front to the WordPress application. This file doesn't do anything, but loads
  * wp-blog-header.php which does and tells WordPress to load the theme.