+++ wp-content/plugins/wordfence/lib/wfAPI.php
@@ -118,9 +118,11 @@ class wfAPI {
 			}
 		}
 		return self::buildQuery(array(
-			'v' => $this->wordpressVersion,
-			's' => $siteurl,
-			'k' => $this->APIKey
+			'v'       => $this->wordpressVersion,
+			's'       => $siteurl,
+			'k'       => $this->APIKey,
+			'openssl' => function_exists('openssl_verify') && defined('OPENSSL_VERSION_NUMBER') ? OPENSSL_VERSION_NUMBER : '0.0.0',
+			'phpv'    => phpversion(),
 		));
 	}