+++ wp-content/plugins/wordfence/lib/wfConfig.php
@@ -507,7 +507,7 @@ class wfConfig {
 		}
 		foreach(self::$securityLevels[2]['otherParams'] as $key => $val){
 			if(isset($_POST[$key])){
-				$ret[$key] = $_POST[$key];
+				$ret[$key] = stripslashes($_POST[$key]);
 			} else {
 				error_log("Missing options param \"$key\" when parsing parameters.");
 			}
@@ -530,7 +530,7 @@ class wfConfig {
 		self::$cache = array();
 	}
 	public static function getHTML($key){
-		return wp_kses(self::get($key), array());
+		return esc_html(self::get($key));
 	}
 	public static function inc($key){
 		$val = self::get($key, false);
@@ -584,16 +584,13 @@ class wfConfig {
 			}
 		}
 
-		if(! isset(self::$cache[$key])){ 
+		if(!array_key_exists($key, self::$cache)){ 
 			$val = self::loadFromDiskCache($key);
 			//$val = self::getDB()->querySingle("select val from " . self::table() . " where name='%s'", $key);
-			if(isset($val)){
-				self::$cache[$key] = $val;
-			} else {
-				self::$cache[$key] = $default;
-			}
+			self::$cache[$key] = $val;
 		}
-		return self::$cache[$key];
+		$val = self::$cache[$key];
+		return $val !== null ? $val : $default;
 	}
 	public static function loadFromDiskCache($key){
 		if(! self::$diskCacheDisabled){