+++ wp-content/themes/looptheme/templates/header.php
@@ -6,15 +6,14 @@
 <img class="secLogo" src="/wp-content/themes/looptheme/assets/images/Awebgo_SpaceShip.png" alt="">
 <div class="dotArea"></div>
 <header class="banner" role="banner">
-  <div class="container">
-    <?php /* <a class="brand" href="<?= esc_url(home_url('/')); ?>"><?php bloginfo('name'); ?></a> */ ?>
-    <a class="brand" href="<?= esc_url(home_url('/')); ?>"><img class="mainLogo hastip" title="[Klick] Gehe zur Startseite" src="/wp-content/themes/looptheme/assets/images/AwebgoLogo.png" alt=""></a>
-    <nav role="navigation">
-      <?php
-      if (has_nav_menu('primary_navigation')) :
-        wp_nav_menu(['theme_location' => 'primary_navigation', 'menu_class' => 'nav']);
-      endif;
-      ?>
-    </nav>
-  </div>
+    <div class="container">
+        <?php /* <a class="brand" href="<?= esc_url(home_url('/')); ?>"><?php bloginfo('name'); ?></a> */ ?>
+        <a class="brand" href="<?= esc_url(home_url('/')); ?>"><img class="mainLogo hastip" title="[Klick] Gehe zur Startseite" src="/wp-content/themes/looptheme/assets/images/AwebgoLogo.png" alt=""></a>
+        <nav role="navigation">
+            <?php
+			if (has_nav_menu('primary_navigation')) : wp_nav_menu(['theme_location' => 'primary_navigation', 'menu_class' => 'nav']);
+			endif;
+            ?>
+        </nav>
+    </div>
 </header>